moving to a proper parsing model (and readable HTML), incomplete
