+++ libraries/hunspell/add.txt
@@ -3,3 +3,4 @@ superitalics
 superbolds
 codebase
 multistyles
+dx
