+++ model/table.py
@@ -9,9 +9,16 @@ from bulletholes.counter import TCounter as Counter
 from elements.elements import Paragraph, OpenFontpost, CloseFontpost, Image
 
 class CellPost(object):
-    def __init__(self, rowspan, colspan):
-        self.rowspan = rowspan
-        self.colspan = colspan
+    def __init__(self, fields):
+        if 'rowspan' in fields:
+            self.rowspan = int(fields['rowspan'])
+        else:
+            self.rowspan = 1
+        if 'colspan' in fields:
+            self.colspan = int(fields['colspan'])
+        else:
+            self.colspan = 1
+    
     def __str__(self):
         return '<td>'
 
@@ -46,10 +53,10 @@ def _bin(data, odelimit, cdelimit):
             continue
 
 class _Table_cell(Atomic_text):
-    def __init__(self, text):
-        Atomic_text.__init__(self, text[1:-1])
-        self.rs = text[0].rowspan
-        self.cs = text[0].colspan
+    def __init__(self, text, rowspan, colspan):
+        Atomic_text.__init__(self, text)
+        self.rs = rowspan
+        self.cs = colspan
     
     def nail(self, i, j):
         self.col = j
@@ -87,11 +94,8 @@ def _row_height(data, r, y):
     return max(cell._SLUGS[-1]['y'] + cell._SLUGS[-1]['leading'] for cell in cells)
 
 class Table(object):
-    def __init__(self, fields, data):
-        self.fields = fields
-        # process table structure
-        # build rows
-        self.data = [list(_Table_cell(C) for C in _bin(R, '<td>', '</td>')) for R in _bin(data, '<tr>', '</tr>')]
+    def __init__(self, L):
+        self.data = [[_Table_cell(C, td.rowspan, td.colspan) for td, C in R] for tr, R in L[1]]
         self._build_matrix()
 
     def flatten(self):
